pretty: move pretty module to scylla-cql
