+++ scylla/src/observability/driver_tracing.rs
@@ -2,11 +2,10 @@ use crate::cluster::node::Node;
 use crate::network::Connection;
 use crate::response::query_result::QueryResult;
 use crate::routing::{Shard, Token};
-use crate::utils::pretty::CqlValueDisplayer;
 use itertools::{Either, Itertools};
 use scylla_cql::frame::response::result::ColumnSpec;
 use scylla_cql::frame::response::result::RawMetadataAndRawRows;
-use scylla_cql::value::deser_cql_value;
+use scylla_cql::value::{deser_cql_value, CqlValueDisplayer};
 use std::borrow::Borrow;
 use std::fmt::Display;
 use std::sync::atomic::AtomicUsize;
