+++ scylla/src/utils/mod.rs
@@ -1,6 +1,4 @@
 pub(crate) mod parse;
 
-pub(crate) mod pretty;
-
 #[cfg(test)]
 pub(crate) mod test_utils;
